怪盗の役職入れ替えの処理の分割
